search in chats completed

+++ server/controllers/chatController.js
@@ -12,6 +12,7 @@ import crypto from "crypto";
 import "../utils/loadEnv.js";
 import createRandomInviteLink from "../utils/createInviteLink.js";
 import indexFinder from "../utils/indexFinder.js";
+import { objectId } from "../utils/typeConverter.js";
 const createChat = async (req, res) => {
   const {
     body: body,
@@ -21,7 +22,7 @@ const createChat = async (req, res) => {
   const data = await Validators.createChat.validate(body);
   // if type of chat is private
   if (data.chatType == chatType[1]) {
-    const chatExists = await Services.findOne('chat',{
+    const chatExists = await Services.findOne("chat", {
       and: [
         { "members.0.memberId": { $in: data.memberIds } },
         { "members.1.memberId": { $in: data.memberIds } },
@@ -59,15 +60,15 @@ const createChat = async (req, res) => {
   });
 
   data["members"] = members;
-  
 
-  const chat = await Services.create(data.chatType,data);
-  await Services.updateMany('user',
+  const chat = await Services.create(data.chatType, data);
+  await Services.updateMany(
+    "user",
     {
       _id: { $in: data.memberIds },
     },
     {
-      $push: {chats:{chatInfo:chat._id} },
+      $push: { chats: { chatInfo: chat._id } },
     }
   );
   await RH.SendResponse({
@@ -85,8 +86,8 @@ const getChat = async (req, res) => {
     user: { userId },
     params: { id: chatId },
   } = req;
-  const user = await Services.findOne('user',{_id:userId})
-  const chat = await Services.findOne('chat',{ _id: chatId });
+  const user = await Services.findOne("user", { _id: userId });
+  const chat = await Services.findOne("chat", { _id: chatId });
   let addedAt;
   user.chats.forEach((chat) => {
     if (chat.chatInfo.equals(chatId)) {
@@ -111,7 +112,7 @@ const getChat = async (req, res) => {
   // });
   const messageIds = chat.messages.map((message) => message.messageInfo);
 
-  const messages = await Services.findMany('message',{
+  const messages = await Services.findMany("message", {
     _id: { $in: messageIds },
     // createdAt: { $gte: joinedAt },
   });
@@ -129,12 +130,10 @@ const getChat = async (req, res) => {
       i++;
       continue;
     }
-    
+
     chat.messages[index].messageInfo = messages[i];
     i++;
   }
-
-  
 
   await RH.SendResponse({
     res,
@@ -151,18 +150,20 @@ const getChats = async (req, res) => {
   const {
     user: { userId },
   } = req;
-  const user = await Services.findOne('user',{_id:userId})
-  const chatIds = user.chats.map((chat)=>chat.chatInfo)
-  const chats = await Services.findMany('chat',
-    { _id: {$in:chatIds} },
+  const user = await Services.findOne("user", { _id: userId });
+  const chatIds = user.chats.map((chat) => chat.chatInfo);
+  const chats = await Services.findMany(
+    "chat",
+    { _id: { $in: chatIds } },
     {},
-    {updatedAt:-1}
+    { updatedAt: -1 }
   );
 
   const messageIds = chats.map(
     (chat) => chat.messages[chat.messages.length - 1]?.messageInfo
   );
-  const messages = await Services.findMany('message',
+  const messages = await Services.findMany(
+    "message",
     {
       _id: { $in: messageIds },
     },
@@ -175,23 +176,23 @@ const getChats = async (req, res) => {
     if (!chat.messages.length) {
       return;
     }
-  
-    chat.messages.splice(0, chat.messages.length - 1);
-    if(!messages[index] || !chat.messages[0].messageInfo.equals(messages[index]._id) ){
-      messages.forEach((message,i)=>{
-        if(message._id.equals(chat.messages[0].messageInfo)){
-          chat.messages[0].messageInfo = messages[i];
 
+    chat.messages.splice(0, chat.messages.length - 1);
+    if (
+      !messages[index] ||
+      !chat.messages[0].messageInfo.equals(messages[index]._id)
+    ) {
+      messages.forEach((message, i) => {
+        if (message._id.equals(chat.messages[0].messageInfo)) {
+          chat.messages[0].messageInfo = messages[i];
         }
-      })
-    }else{
+      });
+    } else {
       chat.messages[0].messageInfo = messages[index];
       index++;
-
     }
-    
   });
-  
+
   // let index;
   // let i = 0;
   // let length = chats.length;
@@ -213,13 +214,6 @@ const getChats = async (req, res) => {
   //   i++;
   // }
 
-
-
-
-
-
-
-
   await RH.SendResponse({
     res,
     statusCode: StatusCodes.OK,
@@ -229,20 +223,20 @@ const getChats = async (req, res) => {
     },
   });
 };
-const addToChats = async (req, res)=>{
-  const {params:{id:chatId}, user:{userId}} = req
+const addToChats = async (req, res) => {
+  const {
+    params: { id: chatId },
+    user: { userId },
+  } = req;
 
   // if chat does not exists
   // error
 
-  await Services.findByIdAndUpdate('user',userId,{
-    $push:{chats:{chatInfo:chatId , addedAt:Date.now()}}
-  })
-  RH.SendResponse({res,statusCode:StatusCodes.OK,title:"ok"})
-
-
-
-}
+  await Services.findByIdAndUpdate("user", userId, {
+    $push: { chats: { chatInfo: chatId, addedAt: Date.now() } },
+  });
+  RH.SendResponse({ res, statusCode: StatusCodes.OK, title: "ok" });
+};
 const pinUnpinChat = async (req, res) => {
   const {
     body,
@@ -276,13 +270,13 @@ const pinUnpinChat = async (req, res) => {
   updateQuery[op]["pinnedChats"] = chatId;
 
   if (data.allChats) {
-    await Services.findByIdAndUpdate('user',userId, updateQuery);
-    await Services.findByIdAndUpdate('chat',chatId, {
+    await Services.findByIdAndUpdate("user", userId, updateQuery);
+    await Services.findByIdAndUpdate("chat", chatId, {
       $set: { pinned: data.pin },
     });
   } else {
     updateQuery["$set"]["chats.$[chat].pinned"] = data.pin;
-    await Services.findByIdAndUpdate('folder',data.folderId, updateQuery, {
+    await Services.findByIdAndUpdate("folder", data.folderId, updateQuery, {
       arrayFilters: [{ "chat.chatInfo": chatId }],
     });
   }
@@ -294,21 +288,18 @@ const pinUnpinChat = async (req, res) => {
 };
 
 const DeleteChat = async (deleteInfo) => {
-  const {
-    chatId ,
-    userId ,
-    deleteAll
-  } = deleteInfo;
-  const chat = await Services.findOne('chat',{ _id: chatId });
-  const user = await Services.findOne('user',{ _id: userId });
+  const { chatId, userId, deleteAll } = deleteInfo;
+  const chat = await Services.findOne("chat", { _id: chatId });
+  const user = await Services.findOne("user", { _id: userId });
   if (deleteAll) {
     if (chat.chatType == "group") {
       if (!chat.owner.equals(userId)) {
-        return res.send("no access")
+        return res.send("no access");
       }
     }
-    await Services.deleteOne('chat',{ _id: chatId });
-    await Services.updateMany('folder',
+    await Services.deleteOne("chat", { _id: chatId });
+    await Services.updateMany(
+      "folder",
       {
         "chats.chatInfo": chatId,
       },
@@ -318,22 +309,24 @@ const DeleteChat = async (deleteInfo) => {
     );
     // pinned chats
     const memberIds = chat.members.map((member) => member.memberId);
-    await Services.updateMany('user',
+    await Services.updateMany(
+      "user",
       {
         _id: { $in: memberIds },
       },
       {
-        $pull: { pinnedChats: chatId , chats:{chatInfo:chatId} },
+        $pull: { pinnedChats: chatId, chats: { chatInfo: chatId } },
       }
     );
   } else {
-    if(chat.chatType == "group"){
-      await Services.findByIdAndUpdate('chat',chatId, {
+    if (chat.chatType == "group") {
+      await Services.findByIdAndUpdate("chat", chatId, {
         $pull: { members: { memberId: userId } },
       });
     }
-    
-    await Services.updateMany('folder',
+
+    await Services.updateMany(
+      "folder",
       {
         _id: { $in: user.folders },
       },
@@ -342,8 +335,8 @@ const DeleteChat = async (deleteInfo) => {
       }
     );
 
-    await Services.findByIdAndUpdate('user',userId, {
-      $pull: { chats:{chatInfo:chatId} ,pinnedChats: chatId },
+    await Services.findByIdAndUpdate("user", userId, {
+      $pull: { chats: { chatInfo: chatId }, pinnedChats: chatId },
     });
   }
 
@@ -355,16 +348,17 @@ const deleteChat = async (req, res) => {
     user: { userId },
     body: { deleteAll },
   } = req;
-  const chat = await Services.findOne('chat',{ _id: chatId });
-  const user = await Services.findOne('user',{ _id: userId });
+  const chat = await Services.findOne("chat", { _id: chatId });
+  const user = await Services.findOne("user", { _id: userId });
   if (deleteAll) {
     if (chat.chatType == "group") {
       if (!chat.owner.equals(userId)) {
-        return res.send("no access")
+        return res.send("no access");
       }
     }
-    await Services.deleteOne('chat',{ _id: chatId });
-    await Services.updateMany('folder',
+    await Services.deleteOne("chat", { _id: chatId });
+    await Services.updateMany(
+      "folder",
       {
         "chats.chatInfo": chatId,
       },
@@ -374,22 +368,24 @@ const deleteChat = async (req, res) => {
     );
     // pinned chats
     const memberIds = chat.members.map((member) => member.memberId);
-    await Services.updateMany('user',
+    await Services.updateMany(
+      "user",
       {
         _id: { $in: memberIds },
       },
       {
-        $pull: { pinnedChats: chatId , chats:{chatInfo:chatId} },
+        $pull: { pinnedChats: chatId, chats: { chatInfo: chatId } },
       }
     );
   } else {
-    if(chat.chatType == "group"){
-      await Services.findByIdAndUpdate('chat',chatId, {
+    if (chat.chatType == "group") {
+      await Services.findByIdAndUpdate("chat", chatId, {
         $pull: { members: { memberId: userId } },
       });
     }
-    
-    await Services.updateMany('folder',
+
+    await Services.updateMany(
+      "folder",
       {
         _id: { $in: user.folders },
       },
@@ -398,12 +394,199 @@ const deleteChat = async (req, res) => {
       }
     );
 
-    await Services.findByIdAndUpdate('user',userId, {
-      $pull: { chats:{chatInfo:chatId} ,pinnedChats: chatId },
+    await Services.findByIdAndUpdate("user", userId, {
+      $pull: { chats: { chatInfo: chatId }, pinnedChats: chatId },
     });
   }
 
   res.send("ok");
 };
 
-export {addToChats,DeleteChat,deleteChat, pinUnpinChat, createChat, getChat, getChats };
+const searchChat = async (req, res) => {
+  const {
+    params: { search },
+    user: { userId },
+  } = req;
+  const user = await Services.findOne("user", { _id: userId });
+  let userChats = user.chats.map((chat) => chat.chatInfo);
+  userChats = await objectId(userChats);
+  // let groupNameRegex1 = `${search}.*`;
+  let groupNameRegex1 = new RegExp(`^${search}`, "i");
+  let groupNameRegex2 = new RegExp(` ${search}`, "i");
+  const result = await Services.aggregate("user", [
+    {
+      $match: { _id: await objectId(userId) },
+    },
+    {
+      $project: {
+        _id: 1,
+        name: 1,
+        contacts: {
+          $filter: {
+            input: "$contacts",
+            as: "contact",
+            cond: {
+              $or: [
+                {
+                  $regexFind: {
+                    input: "$$contact.name",
+                    regex: groupNameRegex1,
+                  },
+                },
+                {
+                  $regexFind: {
+                    input: "$$contact.lastname",
+                    regex: groupNameRegex1,
+                  },
+                },
+              ],
+            },
+          },
+        },
+      },
+    },
+  ]);
+  let contacts = result[0].contacts.map((contact)=>contact.userId)
+  contacts = await objectId(contacts)
+  const chats = await Services.aggregate("chat", [
+    {
+      $match: {
+        _id: { $in: userChats },
+      },
+    },
+    {
+      $addFields: {
+        returnObject1: {
+          $regexFind: { input: "$name", regex: groupNameRegex1 },
+        },
+      },
+    },
+    {
+      $addFields: {
+        returnObject2: {
+          $regexFind: { input: "$name", regex: groupNameRegex2 },
+        },
+      },
+    },
+
+    { $unwind: "$members" },
+    {
+      $lookup: {
+        from: "users",
+        localField: "members.memberId",
+        foreignField: "_id",
+        as: "membersInfo",
+      },
+    },
+    { $unwind: "$membersInfo" },
+    {
+      $addFields: {
+        returnObject3: {
+          $regexFind: { input: "$membersInfo.name", regex: groupNameRegex1 },
+        },
+      },
+    },
+    {
+      $addFields: {
+        returnObject4: {
+          $regexFind: {
+            input: "$membersInfo.lastname",
+            regex: groupNameRegex1,
+          },
+        },
+      },
+    },
+
+    // Group back to arrays
+    {
+      $group: {
+        name: { $first: "$name" },
+        obj1: { $first: "$returnObject1" },
+        obj2: { $first: "$returnObject2" },
+        chatType: { $first: "$chatType" },
+        _id: "$_id",
+        members: { $push: "$members" },
+        membersInfo: { $push: "$membersInfo" },
+      },
+    },
+
+    {
+      $project: {
+        _id: 1,
+        name: 1,
+        chatType: 1,
+        obj1: 1,
+        obj2: 1,
+        membersInfo: {
+          $filter: {
+            input: "$membersInfo",
+            as: "memberInfo",
+            cond: {
+              $and: [
+                {
+                  $or: [
+                    {
+                      $regexFind: {
+                        input: "$$memberInfo.name",
+                        regex: groupNameRegex1,
+                      },
+                    },
+                    {
+                      $regexFind: {
+                        input: "$$memberInfo.lastname",
+                        regex: groupNameRegex1,
+                      },
+                    },
+                    {
+                      $regexFind: {
+                        input: "$$memberInfo.username",
+                        regex: groupNameRegex1,
+                      },
+                    },
+                    {
+                      $in:["$$memberInfo._id",contacts]
+                      
+                    }
+                  ],
+                },
+                { $ne: ["$$memberInfo._id", await objectId(userId)] },
+                { $eq: ["$chatType", "private"] },
+              ],
+            },
+          },
+        },
+      },
+    },
+  ]);
+
+  for (let index = 0; index < chats.length; index++) {
+    if (
+      !chats[index].obj1 &&
+      !chats[index].obj2 &&
+      !chats[index].membersInfo.length
+    ) {
+      chats.splice(index, 1);
+      index--;
+    }
+  }
+  // chats.forEach((chat, index) => {
+  //   if (!chat.obj1 && !chat.obj2 && !chat.membersInfo.length) {
+  //     chats.splice(index, 1);
+  //   }
+  // });
+ 
+  RH.SendResponse({res, statusCode:StatusCodes.OK,title:"ok", value:{
+    chats
+  }})
+};
+
+export {
+  addToChats,
+  DeleteChat,
+  deleteChat,
+  pinUnpinChat,
+  createChat,
+  getChat,
+  getChats,
+  searchChat,
+};
